Styled excercise search page
